[V9990] Fixed V9_Peek16_CurrentAddr() content and prototype

+++ engine/src/v9990.c
@@ -317,7 +317,7 @@ u8 V9_Peek_CurrentAddr() __PRESERVES(b, c, d, e, h, l, iyl, iyh)
 
 //-----------------------------------------------------------------------------
 //
-u8 V9_Peek16_CurrentAddr() __PRESERVES(b, c, h, l, iyl, iyh)
+u16 V9_Peek16_CurrentAddr() __PRESERVES(b, c, h, l, iyl, iyh)
 {
 	__asm
 		in		a, (V9_P00)
